chat-with-csv/requirements.txt
pandasai
streamlit
matplotlib
chainlit
python-dotenv
pyyaml